refactor: reorganize chart utilities and update import paths
- Moved chart utility functions to a new directory structure for better organization.
- Updated import paths in various components to reflect the new structure.
- Removed unused chart-related files to streamline the codebase.
