+++ src/ui/charts/options.ts
@@ -1,4 +1,4 @@
-import { EChartsOption } from 'echarts'
+import type { EChartsOption } from 'echarts'
 
 export const defaultLineChartOption: EChartsOption = {
   backgroundColor: 'transparent',
